clean up, improve architecture
