+++ app/repositories/polls.ts
@@ -1,12 +1,11 @@
 import { sql, type Transaction } from 'kysely'
+import { getArtefact } from '#app/repositories/artefact.ts'
 import {
 	createPost,
 	getDescendantCount,
 	getPost,
 } from '#app/repositories/post.ts'
 import {
-	type Quote,
-	type Artefact,
 	type Claim,
 	type PollType,
 	type Post,
@@ -17,96 +16,7 @@ import {
 	getCandidateClaim,
 	updatePostIdOnCandidateClaim,
 } from './claim-extraction.ts'
-
-export async function getOrCreateArtefact(
-	trx: Transaction<DB>,
-	url: string,
-): Promise<Artefact> {
-	const existingArtefact: Artefact | undefined = await trx
-		.selectFrom('Artefact')
-		.where('url', '=', url)
-		.selectAll()
-		.executeTakeFirst()
-
-	if (existingArtefact !== undefined) {
-		return existingArtefact
-	}
-
-	const createdArtefact = await trx
-		.insertInto('Artefact')
-		.values({
-			url: url,
-		})
-		.returningAll()
-		.executeTakeFirstOrThrow()
-
-	return {
-		id: createdArtefact.id,
-		url: createdArtefact.url,
-		createdAt: createdArtefact.createdAt,
-	}
-}
-
-export async function getArtefact(
-	trx: Transaction<DB>,
-	id: number,
-): Promise<Artefact> {
-	const result = await trx
-		.selectFrom('Artefact')
-		.where('id', '=', id)
-		.selectAll()
-		.executeTakeFirstOrThrow()
-
-	return {
-		id: result.id,
-		url: result.url,
-		createdAt: result.createdAt,
-	}
-}
-
-export async function getOrCreateQuote(
-	trx: Transaction<DB>,
-	artefactId: number,
-	quote: string,
-): Promise<Quote> {
-	const existingQuote: Quote | undefined = await trx
-		.selectFrom('Quote')
-		.where('quote', '=', quote)
-		.where('artefactId', '=', artefactId)
-		.selectAll()
-		.executeTakeFirst()
-
-	if (existingQuote !== undefined) {
-		return existingQuote
-	}
-
-	return await trx
-		.insertInto('Quote')
-		.values({
-			artefactId: artefactId,
-			quote: quote,
-		})
-		.returningAll()
-		.executeTakeFirstOrThrow()
-}
-
-export async function getQuote(
-	trx: Transaction<DB>,
-	id: number,
-): Promise<Quote> {
-	const result = await trx
-		.selectFrom('Quote')
-		.where('id', '=', id)
-		.selectAll()
-		.executeTakeFirstOrThrow()
-
-	return {
-		id: result.id,
-		artefactId: result.artefactId,
-		quote: result.quote,
-		createdAt: result.createdAt,
-	}
-}
+import { getQuote } from './quote.ts'
 
 export async function createClaim(
 	trx: Transaction<DB>,
