+++ app/routes/artefact.$artefactId.quote.$quoteId.tsx
@@ -13,9 +13,11 @@ import {
 	TabsTrigger,
 } from '#app/components/ui/tabs.tsx'
 import { db } from '#app/db.ts'
+import { getArtefact } from '#app/repositories/artefact.ts'
 import { getCandidateClaims } from '#app/repositories/claim-extraction.ts'
 import { getQuoteFallacies } from '#app/repositories/fallacy-detection.ts'
-import { getArtefact, getPollPost, getQuote } from '#app/repositories/polls.ts'
+import { getPollPost } from '#app/repositories/polls.ts'
+import { getQuote } from '#app/repositories/quote.ts'
 import {
 	type Artefact,
 	PollType,
